moved entrybyplaintext to bibtexextractor package
