+++ src/main/java/org/jabref/gui/bibtexextractor/FailedToParseDialog.fxml
@@ -2,8 +2,7 @@
 
 
 <?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
+<?import javafx.scene.layout.VBox?>
 <DialogPane prefHeight="430.0" prefWidth="586.0" xmlns="http://javafx.com/javafx/8.0.171"
             xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.entrybyplaintext.FailedToParseDialog">
     <content>
